feat: write openai methods as service pattern
