+++ src/app/api/daily-summary/route.ts
@@ -7,8 +7,6 @@ import {
 import type { AISummaryData } from '@/shared/types';
 import {
   handleApiRequest,
-  callOpenAIAndParseJSON,
-  validateAISummaryStructure,
   type ApiContext,
   withValidation,
   VALIDATION_SCHEMAS,
@@ -83,6 +81,7 @@ export async function POST(request: NextRequest) {
         const dailySummaryService = ServiceFactory.createDailySummaryService(
           context.supabase
         );
+        const openAIService = ServiceFactory.createOpenAIService();
 
         const prompt = getAISummaryPrompt(
           validatedData.locale as Locale,
@@ -92,23 +91,27 @@ export async function POST(request: NextRequest) {
           validatedData.locale as Locale
         );
 
-        const summary = await callOpenAIAndParseJSON(
-          [
+        const summary = await openAIService.callOpenAIAndParseJSON({
+          messages: [
             { role: 'system', content: systemPrompt },
             { role: 'user', content: prompt },
           ],
-          {
+          maxTokens: 4000,
+          options: {
             span: context.span,
             operation: 'create_daily_summary',
-            maxTokens: 4000,
-          }
-        );
+          },
+        });
 
-        const validatedSummary = validateAISummaryStructure(
+        const validatedSummary = openAIService.validateAISummaryStructure({
           summary,
-          AI_SUMMARY_REQUIRED_FIELDS,
-          AI_SUMMARY_ARRAY_FIELDS
-        ) as unknown as AISummaryData;
+          requiredFields: AI_SUMMARY_REQUIRED_FIELDS,
+          arrayFields: AI_SUMMARY_ARRAY_FIELDS,
+          options: {
+            span: context.span,
+            operation: 'create_daily_summary',
+          },
+        }) as unknown as AISummaryData;
 
         await dailySummaryService.saveSummary(
           validatedSummary,
