+++ src/shared/lib/api/index.ts
@@ -1,4 +1,3 @@
-export * from './openai';
 export * from './services';
 export * from './errors';
 export * from './middleware';
