+++ src/shared/lib/api/services/index.ts
@@ -3,3 +3,4 @@ export * from './weekly-summary.service';
 export * from './notes.service';
 export * from './feedback.service';
 export * from './speech-to-text.service';
+export * from './openai.service';
